more fixes

+++ core/include/core/ppu.h
@@ -49,7 +49,7 @@ uint8_t ppu_get_pixel(ppu_t* ppu, size_t x, size_t y);
 uint8_t ppu_get_tile(ppu_t* ppu, mmu_t* mmu, uint8_t tile_id, size_t tile_x, size_t tile_y, bool is_sprite);
 uint8_t ppu_convert_palette(ppu_t* ppu, uint8_t pixel, uint8_t palette);
 
-uint8_t ppu_render_background(ppu_t* ppu, mmu_t* mmu, size_t x, size_t y, uint8_t is_window);
+uint8_t ppu_render_background(ppu_t* ppu, mmu_t* mmu, uint8_t x, uint8_t y, uint8_t is_window);
 void ppu_render_sprites(ppu_t* ppu, mmu_t* mmu, size_t x, size_t y, uint8_t background);
 void ppu_render_line(ppu_t* ppu, mmu_t* mmu, size_t line);
